Update high_resolution_filament_sensor.py to new klipper extruder.move syntax
For one of the newer klipper updates, some syntax was changed that broke the code for this sensor. 

_capture_extruder_move now requires 4 inputs so I added "ea_index" from the klipper toolhead.py script. I don't think its used/needed for the sensor code, but it takes the input anyway just so it doesn't make klipper throw an error.

extruder.move() was changed to extruder.process_move() so I updated the sections that used that to the new syntax.

This was tested incrementally until I was able to run a print without an error. Originally the printer wasn't able to turn on. That was fixed with changing extruder.move to extruder.process_move. Afterwards, the printer would error at the start of new prints due to various things like flush_handler, pressure_advance errors, etc.